+++ Tests/AccessibilityTests/AccessibilityProject/Sources/ExternalTarget/ExternalProtocol.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol ExternalProtocol {
+    func someExternalProtocolMethod()
+}