+++ Tests/AccessibilityTests/AccessibilityProject/Sources/TargetA/PublicClassAdoptingExternalProtocolObjcAccessible.swift
@@ -1,0 +1,8 @@
+import Foundation
+import ExternalTarget
+
+// Not referenced in MainTarget but retained due to retainObjcAccessible.
+@objc public class PublicClassAdoptingExternalProtocolObjcAccessible: NSObject, ExternalProtocol {
+    public override init() {}
+    public func someExternalProtocolMethod() {}
+}